+++ lua/legendary/scratchpad.lua
